Refactor consultant profile components and add availability and pricing sections
